test: Add test to verify cleanup of stale asset references upon asset deletion
